Split save-dialog accessory regression test into its own wired test file
CmuxConfigActionSaverTests.swift crossed the 500-line untracked budget
threshold; the dialog-accessory disclosure regression test now lives in
WorkspaceActionSaveDialogAccessoryTests.swift, wired into the cmuxTests
target (build file, file reference, group child, sources phase).
