+++ cmuxTests/CmuxConfigActionSaverTests.swift
@@ -1,4 +1,3 @@
-import AppKit
 import Foundation
 import Testing
 
@@ -444,47 +443,6 @@ struct CmuxConfigActionSaverTests {
         #expect(plain.capturedCommands == [])
         #expect(plain.capturedURLs == [])
         #expect(plain.capturedEnvironmentKeys == [])
-    }
-
-    @Test @MainActor func saveDialogDisclosureTextViewsSizeToTheirContent() {
-        // Regression: the disclosure NSTextViews were created 1pt tall and
-        // NSText clamps vertical growth to maxSize (which defaults to the
-        // initial frame size), so the sensitive commands/URLs/env text the
-        // dialog exists to show rendered blank inside its scroll view.
-        let accessory = WorkspaceActionSaveDialogAccessory(
-            snapshot: WorkspaceConfigActionSnapshot(
-                definition: CmuxWorkspaceDefinition(
-                    name: "W",
-                    env: ["API_TOKEN": "secret"],
-                    layout: .pane(CmuxPaneDefinition(surfaces: [
-                        CmuxSurfaceDefinition(type: .terminal, command: "claude --model claude-fable-5"),
-                        CmuxSurfaceDefinition(type: .browser, url: "https://example.com/callback?code=abc"),
-                    ]))
-                ),
-                skippedPanelCount: 0
-            ),
-            initialName: "Layout"
-        )
-
-        func scrollViews(in view: NSView) -> [NSScrollView] {
-            view.subviews.flatMap { subview -> [NSScrollView] in
-                if let scrollView = subview as? NSScrollView { return [scrollView] }
-                return scrollViews(in: subview)
-            }
-        }
-
-        let disclosureScrollViews = scrollViews(in: accessory.view)
-        #expect(disclosureScrollViews.count == 3)
-        for scrollView in disclosureScrollViews {
-            let textView = scrollView.documentView as? NSTextView
-            #expect(textView != nil)
-            guard let textView else { continue }
-            #expect(!textView.string.isEmpty)
-            // The document view must have grown to fit its laid-out text; a
-            // 1pt-tall document view means the disclosure renders blank.
-            #expect(textView.frame.height >= 20)
-            #expect(textView.maxSize.height >= textView.frame.height)
-        }
     }
 
     @Test func snapshotReportsOversizedCommands() {
